Config changes for 9070XT support and new PC

+++ files/dotfiles/Configs/scripts/fix-vaapi-chromium-flatpak.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+set -euxo pipefail
 DEFAULT_APP="com.brave.Browser"
 
 show_usage() {
@@ -12,16 +13,39 @@ EOF
 	exit 1
 }
 
+detect_gpu() {
+	local base_dir="/usr/lib64/dri"
+	! [ -d "/usr/lib64/dri" ] && base_dir="/usr/lib/dri" # fallback
+	local drivers=("radeonsi_drv_video.so" "nvidia_drv_video.so")
+
+	for driver in "${drivers[@]}"; do
+		local path="$base_dir/$driver"
+		if [ -r "$path" ]; then
+			echo "$path"
+			return 0
+		fi
+	done
+	echo ""
+}
+
 do_overrides() {
 	local app=$1
-	local dri=$2
-	local user_opt=$3
+	local local_dri=$2
+	local remote_dri=$3
+	local user_opt=$4
 
-	flatpak override "$user_opt" --reset "$APP" 2>/dev/null
+	flatpak override "$user_opt" --reset "$app" 2>/dev/null
+
+	# Set LIBVA_DRIVER_NAME based on GPU type
+	local libva_driver=()
+	if [[ "$local_dri" == *nvidia* ]]; then
+		libva_driver=("--env=LIBVA_DRIVER_NAME=nvidia --env=NVD_BACKEND=direct")
+	fi
+
 	flatpak override "$user_opt" \
-		--env=LIBVA_DRIVER_NAME=nvidia \
-		--env=LIBVA_DRIVERS_PATH="$dri" \
-		--env=NVD_BACKEND=direct "$app"
+		--env=LIBVA_DRIVERS_PATH="$(dirname "$remote_dri")" \
+		"${libva_driver[@]}" \
+		"$app"
 }
 
 USER_FLAG=false
@@ -68,10 +92,10 @@ while [ "$#" -gt 0 ]; do
 done
 
 APP="${APP_NAME:-$DEFAULT_APP}"
-LOCAL_DRI_PATH="${LOCAL_PATH:-/usr/lib64/dri/nvidia_drv_video.so}"
+LOCAL_DRI_PATH="${LOCAL_PATH:-$(detect_gpu)}"
 BASE_DIR="$(realpath "$HOME")/.var/app/$APP"
 DRI_PATH="$BASE_DIR/dri"
-REMOTE_DRI_PATH="$DRI_PATH/nvidia_drv_video.so"
+REMOTE_DRI_PATH="$DRI_PATH/$(basename "$LOCAL_DRI_PATH")"
 user_opt="${USER_FLAG:+--user}"
 
 [ -z "$APP_NAME" ] && APP_NAME=$APP
@@ -81,18 +105,21 @@ if ! (flatpak "$user_opt" list | grep -q "$APP"); then
 	exit 1
 fi
 
+if ! [ -r "$LOCAL_DRI_PATH" ]; then
+	echo "Unable to detect a supported VAAPI GPU driver, exiting..."
+	exit 1
+fi
+
 if [ -r "$REMOTE_DRI_PATH" ]; then
 	do_overrides "$APP" "$REMOTE_DRI_PATH" "$user_opt"
 	echo "Detected '$REMOTE_DRI_PATH', fix has already run, redoing overrides..."
 	exit 0
 fi
 
-[ ! -r "$LOCAL_DRI_PATH" ] && echo "Can't find driver at $LOCAL_DRI_PATH" && exit 1
-
 rm -rf "$DRI_PATH" &&
 	mkdir -p "$DRI_PATH" &&
-	cp -f "$LOCAL_DRI_PATH" "$DRI_PATH/nvidia_drv_video.so"
+	cp -f "$LOCAL_DRI_PATH" "$REMOTE_DRI_PATH"
 
-do_overrides "$APP" "$REMOTE_DRI_PATH" "$user_opt"
+do_overrides "$APP" "$LOCAL_DRI_PATH" "$REMOTE_DRI_PATH" "$user_opt"
 
-echo "NVIDIA driver configured for $APP at $DRI_PATH"
+echo "VAAPI GPU driver configured for $APP at $DRI_PATH"
